pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "palooza"
description = "Rebootapalooza manager"
authors = [{ name = "Tjaart van der Walt", email = "tjaart@ucsb.edu" }]
dynamic = ["version"]
dependencies = ["flask", "flask-cors", "Flask-SQLAlchemy", "Flask-DB", "sqlalchemy-serializer"]

[project.scripts]
palooza = "palooza:main"
palooza_seed = "palooza:seed"
